move enum validators
